channel and counter becomes integers

+++ libk8055/k8055.h
@@ -38,26 +38,26 @@ extern "C" {
     long search_devices( void );
     int open_device( struct k8055_dev* dev, int board_address );
     int close_device( struct k8055_dev* dev );
-    long read_analog_channel( struct k8055_dev* dev, long channel );
+    long read_analog_channel( struct k8055_dev* dev, int channel );
     int read_all_analog( struct k8055_dev* dev, long* data1, long* data2 );
-    int output_analog_channel( struct k8055_dev* dev ,long channel, long data );
+    int output_analog_channel( struct k8055_dev* dev ,int channel, long data );
     int output_all_analog( struct k8055_dev* dev, long data1, long data2 );
     int clear_all_analog( struct k8055_dev* dev );
-    int clear_analog_channel( struct k8055_dev* dev, long channel );
-    int set_analog_channel( struct k8055_dev* dev, long channel );
+    int clear_analog_channel( struct k8055_dev* dev, int channel );
+    int set_analog_channel( struct k8055_dev* dev, int channel );
     int set_all_analog( struct k8055_dev* dev );
     int write_all_digital( struct k8055_dev* dev, long data );
-    int clear_digital_channel( struct k8055_dev* dev, long channel );
+    int clear_digital_channel( struct k8055_dev* dev, int channel );
     int clear_all_digital( struct k8055_dev* dev );
-    int set_digital_channel( struct k8055_dev* dev, long channel );
+    int set_digital_channel( struct k8055_dev* dev, int channel );
     int set_all_digital( struct k8055_dev* dev );
-    int read_digital_channel( struct k8055_dev* dev, long channel );
+    int read_digital_channel( struct k8055_dev* dev, int channel );
     long read_all_digital( struct k8055_dev* dev );
     int read_all_values( struct k8055_dev* dev, long int* data1, long int* data2, long int* data3, long int* data4, long int* data5 );
     int set_all_values( struct k8055_dev* dev, int digital_data, int ad_data1, int ad_data2 );
-    int reset_counter( struct k8055_dev* dev, long counter );
-    long read_counter( struct k8055_dev* dev, long counter );
-    int set_counter_debounce_time( struct k8055_dev* dev, long counter, long debounce_time );
+    int reset_counter( struct k8055_dev* dev, int counter );
+    long read_counter( struct k8055_dev* dev, int counter );
+    int set_counter_debounce_time( struct k8055_dev* dev, int counter, long debounce_time );
 
     /* Velleman API */
     char* Version( void );
